Fix case when subscription table name is not present or nothing has to be cleaned up

+++ src/NServiceBus.Transport.SqlServer.AcceptanceTests/ConfigureEndpointSqlServerTransport.cs
@@ -61,12 +61,16 @@ public class ConfigureEndpointSqlServerTransport : IConfigureEndpointTestExecuti
                 commandTextBuilder.AppendLine($"IF OBJECT_ID('{delayedAddressQualifiedName}', 'U') IS NOT NULL DROP TABLE {delayedAddressQualifiedName}");
             }
 
-            if (!doNotCleanNativeSubscriptions)
+            if (!doNotCleanNativeSubscriptions && !string.IsNullOrEmpty(subscriptionTableQuotedQualifiedName))
             {
                 commandTextBuilder.AppendLine($"IF OBJECT_ID('{subscriptionTableQuotedQualifiedName}', 'U') IS NOT NULL DROP TABLE {subscriptionTableQuotedQualifiedName}");
             }
 
-            await TryDeleteTables(conn, commandTextBuilder.ToString());
+            var commandText = commandTextBuilder.ToString();
+            if (!string.IsNullOrEmpty(commandText))
+            {
+                await TryDeleteTables(conn, commandText);
+            }
         }
     }
